+++ R/peaks-functions.R
@@ -562,3 +562,45 @@ joinPeaksGnps <- function(x, y, xPrecursorMz = NA_real_,
         x[idx, , drop = FALSE]
     }
 }
+
+#' Combine peaks within each peak matrix if the difference of their m/z is
+#' smaller than `tolerance` and `ppm`. Peak grouping is performed with the
+#' `MsCoreUtils::group` function.
+#'
+#' If peak annotations are present, i.e. `x` has more than 2 columns: peak
+#' annotations from the **first** peak in each group are used.
+#'
+#' Note that `weighted = TRUE` overrides `mzFun` using the `weighted.mean`
+#' to calculate the aggregated m/z values.
+#'
+#' @author Johannes Rainer
+#'
+#' @importFrom stats weighted.mean
+#'
+#' @noRd
+.peaks_combine <- function(x, ppm = 20, tolerance = 0,
+                           intensityFun = base::mean, mzFun = base::mean,
+                           weighted = TRUE, spectrumMsLevel,
+                           msLevel = spectrumMsLevel, ...) {
+    if (!spectrumMsLevel %in% msLevel || !length(x))
+        return(x)
+    grps <- group(x[, "mz"], tolerance = tolerance, ppm = ppm)
+    lg <- length(grps)
+    if (grps[lg] == lg)
+        return(x)
+    mzs <- split(x[, "mz"], grps)
+    ints <- split(x[, "intensity"], grps)
+    if (weighted)
+        mzs <- unlist(mapply(
+            mzs, ints, FUN = function(m, i) weighted.mean(m, i + 1),
+            SIMPLIFY = FALSE, USE.NAMES = FALSE), use.names = FALSE)
+    else mzs <- vapply1d(mzs, mzFun)
+    ints <- vapply1d(ints, intensityFun)
+    if (ncol(x) > 2) {
+        xnew <- x[match(seq_len(grps[lg]), grps), , drop = FALSE]
+        xnew[, "mz"] <- mzs
+        xnew[, "intensity"] <- ints
+        xnew
+    } else
+        cbind(mz = mzs, intensity = ints)
+}